requirements.txt
numpy==1.22.3
Pillow==9.1.0
rsa==4.8
